+++ src/main/java/br/ufrn/imd/app/jsf/LoginBean.java
@@ -1,11 +1,26 @@
 package br.ufrn.imd.app.jsf;
 
+import br.ufrn.imd.app.model.User;
+
 import javax.enterprise.context.RequestScoped;
+import javax.inject.Inject;
 import javax.inject.Named;
 
+/**
+ * Bean that representes the state of login.
+ */
 @Named
 @RequestScoped
 public class LoginBean {
+
+    private static final String LOGIN_PAGE = "login";
+
+    @Inject
+    private UserBean userBean;
+
+    @Inject
+    private MessageBean messageBean;
+
     private String username;
     private String password;
 
@@ -26,6 +41,20 @@ public class LoginBean {
     }
 
     public String login() {
-        return "index";
+        if (username == null || username.isEmpty()) {
+            messageBean.setError("Login: campo obrigatório.");
+            return loginPage();
+        }
+        User user = new User(username);
+        userBean.login(user);
+
+
+        messageBean.setSuccess("Login realizado com sucesso.");
+
+        return "home";
+    }
+
+    public String loginPage() {
+        return LOGIN_PAGE;
     }
 }