+++ src/main/java/br/ufrn/imd/app/jsf/ServiceBean.java
@@ -1,0 +1,19 @@
+package br.ufrn.imd.app.jsf;
+
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Named;
+
+@Named
+@RequestScoped
+public class ServiceBean {
+
+    private String exemplo = "0";
+
+    public String getExemplo() {
+        return exemplo;
+    }
+
+    public void setExemplo(String exemplo) {
+        this.exemplo = exemplo;
+    }
+}