+++ Changelog.md
@@ -3,5 +3,7 @@
 ## Changelog
 
 ### Recent Changes
+- Facade login/signup view
+- Home with service selection
 
 ## RELEASES